Use a new frame for presentations. Add a margin.
Using a frame simplifies a lot of the save/restore logic for hiding lots
of the emacs chrome. Instead we just modify the new frame and then throw
it away when we're done.

Using a new frame also helps with the margins. Although it is possible
to modify the margins of an existing frame this makes the frame bigger
but the window manager doesn't notice. This is particularly problematic
with full screen windows where some of the content is pushed off the
screen.